# Plan/client/swagger-docs-3.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "NoteEarly API Documentation (Zod-Generated)",
      "version": "1.0.0",
      "description": "API documentation for the NoteEarly platform (schemas generated from Zod)."
    },
    "servers": [
      {
        "url": "http://localhost:4000/api/v1",
        "description": "Development server (V1)"
      }
    ],
    "components": {
...
      "schemas": {
...
        "CreateModuleInput": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "structuredContent": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "index": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 5000
                  }
                },
                "required": [
                  "index",
                  "text"
                ]
              },
              "minItems": 1,
              "maxItems": 40
            },
            "level": {
              "type": "integer",
              "enum": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
            },
            "genre": {
              "type": "string",
              "enum": [
                "History",
                "Adventure",
                "Science",
                "Non-Fiction",
                "Fantasy",
                "Biography",
                "Mystery",
                "Science-Fiction",
                "Folktale",
                "Custom"
              ]
            },
            "language": {
              "type": "string",
              "enum": [
                "UK",
                "US"
              ]
            },
            "description": {
              "type": "string",
              "nullable": true,
              "maxLength": 2000
            },
            "imageUrl": {
              "type": "string",
              "nullable": true,
              "maxLength": 2048,
              "format": "uri"
            },
            "estimatedReadingTime": {
              "type": "integer",
              "nullable": true,
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 1440
            },
            "isActive": {
              "type": "boolean"
            }
          },
          "required": [
            "title",
            "structuredContent",
            "level",
            "genre",
            "language"
          ]
        },
...
        "Paragraph": {
          "type": "object",
          "properties": {
            "index": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true
            },
            "text": {
              "type": "string",
              "minLength": 1
            }
          },
          "required": [
            "index",
            "text"
          ]
        },
...
        "ReadingModuleDTO": {
          "description": "Data Transfer Object for Reading Modules",
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "adminId": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            },
            "title": {
              "type": "string"
            },
            "structuredContent": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Paragraph"
              }
            },
            "paragraphCount": {
              "type": "integer"
            },
            "level": {
              "type": "integer",
              "enum": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
            },
            "type": {
              "type": "string",
              "enum": [
                "curated",
                "custom"
              ]
            },
            "genre": {
              "type": "string",
              "enum": [
                "History",
                "Adventure",
                "Science",
                "Non-Fiction",
                "Fantasy",
                "Biography",
                "Mystery",
                "Science-Fiction",
                "Folktale",
                "Custom"
              ]
            },
            "language": {
              "type": "string",
              "enum": [
                "UK",
                "US"
              ]
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "imageUrl": {
              "type": "string",
              "format": "url",
              "nullable": true
            },
            "estimatedReadingTime": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          "required": [
            "id",
            "title",
            "structuredContent",
            "paragraphCount",
            "level",
            "type",
            "genre",
            "language",
            "createdAt"
          ]
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "Enter JWT token **_only_**"
        }
      },
...
    },
    "paths": {
...
      "/analytics/subscriptions": {
        "get": {
          "tags": [
            "Analytics - SuperAdmin"
          ],
          "summary": "Get subscription statistics",
          "description": "Retrieves statistics related to user subscriptions (e.g., counts per tier, revenue).",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Subscription statistics retrieved successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "example": {
                      "totalActive": 120,
                      "freeTier": 30,
                      "homeTier": 50,
                      "proTier": 40,
                      "monthlyRevenue": 599.5
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Unauthorized (not a Super Admin)."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
...
      "/api/v1/reading-modules": {
        "post": {
          "tags": [
            "Reading Modules - Admin"
          ],
          "summary": "Create a new custom reading module (Admin/SuperAdmin)",
          "description": "Allows an authenticated admin or superadmin to create a new custom reading module.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateModuleInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Custom module created successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReadingModuleDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Validation Error."
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Forbidden (e.g., Subscription limit reached)."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
...
      "/api/v1/subscriptions/stripe-webhook": {
        "post": {
          "tags": [
            "Subscriptions - Webhook"
          ],
          "summary": "Handle Stripe webhook events (Public)",
          "description": "Receives events from Stripe (e.g., checkout completed, subscription updated/canceled) and updates the application state accordingly. Access is public but requires a valid Stripe signature.",
          "requestBody": {
            "required": true,
            "description": "Raw event payload from Stripe with signature in header.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Generic Stripe event object structure."
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Webhook received and processed successfully (or acknowledged for later processing).",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "received": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request (e.g., invalid signature, missing Stripe-Signature header, malformed payload)."
            },
            "500": {
              "description": "Internal server error during webhook processing."
            }
          }
        }
      }
    },
...
  }